+++ grid2op/Backend/__init__.py
@@ -1,4 +1,7 @@
-__all__ = ["Backend", "PandaPowerBackend"]
+__all__ = ["Backend",
+           "PandaPowerBackend",
+           "PandaPowerBackendWithDetailedTopo"]
 
 from grid2op.Backend.backend import Backend
 from grid2op.Backend.pandaPowerBackend import PandaPowerBackend
+from grid2op.Backend.pandapowerBackend_with_detailed_topo import PandaPowerBackendWithDetailedTopo